Uploaded Camera html
This is a trial run. I have the video stream working. 
I store it in a <canvas> onclick this displays it to the left of the video. Then I turn into a data url using the toDataURL() javascript function.
This is stored in an <img> tag with the id ="dataU". This is displayed below the video stream.
Then I run a function that passes the image to a URL but first I turn it into a blob using the function makeblob() (thanks stackoverflow)
Then I call call the api using the blob(because the api doesn't take data url types, which is a pain)
I then print the results on the console. 
In future updates I want to lessen the amount of clutter 
!!!!!DO THIS FIRST!!!!! by making only one image box show up rather than 3
!!!!!DO THIS SECOND!!!!! Then, I'll make it so I can read the specific emotions rather than just displaying it on the console, because thats lame

Then hopefully I can implement the weather api
I need to do some database stuff which is a big deal. Talk about it in the future
Then I'll write the algorithm